[CHORE] : Unit tests for Model (#153)
* [CHORE] : Unit tests for Model

This is a first pass to add unit tests for the data model.
It is based on testing the main model objects (like responses from person, petitions, and donations requests) against static JSON objects obtained from the API endpoints.
I have not included too many assertions just in case this isn't the way we want to go. New assertions can easily be added later.

Note : the Search API test was disabled since this is out of the scope for the current version. However, since I already had it in my code and it will be useful for v2, I feel like it would be good to include it.

* [CHORE] : Unit tests for Model (new iteration with asset bundle)

* Clean up of the date formatter since we don't need it anymore with `x-dates-epoch` pulling dates as time interval from the API endpoints.

* Added comments and cleaned up unnecessary comments and methods from the initial template.