Remove Korean from localise.biz (#264)
We reached a limit of 10 languages. I think we can remove Korean because it's not finished and unused, and there's no one that can work on it.  
I'm leaving the .strings files in the repo in case we want to reintroduce the language later - they could be reuploaded to localise.biz.